requirements.txt
Flask==3.0.0
gunicorn==21.2.0
flask-login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Limiter==3.5.0
bcrypt==4.1.2
supabase>=2.24.0
python-dotenv==1.0.0
httpx>=0.26.0
websockets>=15.0.0
requests==2.31.0
urllib3>=2.0.0
bleach==6.1.0

